System.KSP.Auxiliary

Description

Documentation

takeUntil :: (a -> Bool) -> [a] -> [a]

The function takeUntil takes a test function and a list of values. It returns all elements that do not pass the test function until it reaches the first one that passes the test function, otherwise it returns the empty list.

firstJust :: [Maybe a] -> Maybe a

firstJust returns of a list of Maybe values the first actual Just value or nothing if none is within the list.