System.KSP.DataConstructors

Description

Documentation

data Object

An Object is something named.

Constructors

Object

Fields

name :: String

data Celestial

The Celestial type defines a celestial object within this library. Every celestial object has

r [m] the radius (from the bodys center) in meter

m [kg] the mass of the object in kilo gramms

soi [m] the sphere of influence of the object in meter

system empty or a system of surronding objects.

Constructors

Celestial

Fields

r :: Double

m :: Double

soi :: Double

system :: System Body Body

data Body

The Body type is the central data type within the KSP library. It could be either Railed or Movable

Railed is everything withing ksp that is a star, a planet or a moon. They are static in their orbits and are therefore called railed.

Movable is every other part in ksp, that is physical. Here they include deltaV counter, but are also modelled as celestial objects. Actually in ksp orbits around movable parts arent possible. Anyhow, within this data type such constructs are doable.

Constructors

Railed

Fields

object :: Object

celestial :: Celestial

Movable

Fields

object :: Object

celestial :: Celestial

deltaV :: DeltaV

data Orbit a

The Orbit module contains the type definition of a Orbit.

Constructors

Landed

O

Fields

centerBody :: a

apoapsis :: Double

periapsis :: Double

inclination :: Maybe Double

omega_big :: Maybe Double

omega_small :: Maybe Double

data System b a

The System module contains the type definition of a System.

Constructors

Empty

System [(Orbit b, a)]

type GravConst = Double

GravConst is the type of the gravitation constant, which is obious a double.

type KSystem = System Body

KSystem creates a data constructor, that binds the System to a System of Orbits around Bodys.