System.KSP.DataDestructors

Description

Documentation

getNextUp :: KSystem Body -> Body -> Maybe Body

getNextUp takes a KSystem system and a Body. It returns the body that the supplied body orbits. Nothing is returned if the supplied body is not found within the system.

getPathUp :: KSystem Body -> Body -> [Body]

getPathUp takes a KSystem system and a Body. It returns the chain of body's that are orbiting each other.

getDivid :: KSystem Body -> Body -> Body -> (Int, Body)

getDivid takes a KSystem system and a from Body and a to Body. For both bodys, the path up is calculated by getPathUp and the position, where they reach the same body is returned (as Position and with the actual body).

sOrbitInSystem :: Body -> KSystem Body -> Maybe (Orbit Body)

sOrbitInSystem takes a body, a KSystem system and maybe returns the orbit of the body.

pathOBetween :: KSystem Body -> Body -> Body -> [(Body, Orbit Body)]

pathOBetween takes a KSystem system and a from Body and a to Body. If one want's to reach the to body from from body the bodys between and their orbit around their centerBody are returnd.

pathBetween_ :: KSystem Body -> Body -> Body -> [Body]

pathBetween_ takes a KSystem system and a from Body and a to Body. If one want's to reach the to body from from body the bodys between are returnd.

pathBetween' :: KSystem Body -> Body -> Body -> ([Body], Body, [Body])

pathBetween takes a KSystem system and from Body and a to Body. If one want's to reach the to body from from body the bodys between and their orbit around their centerBody are returnd. They are returnd as triple, with the overlapping body in the center of the triple and the path of the from body in the first element of the triple. The path of the to body in the last element of the triple.

speeds :: [(Body, Orbit Body)] -> [Speed]

speeds takes the result of pathOBetween and calculates the corresponding orbital speeds.