System.KSP.OrbitalMechanics

Description

Documentation

type Radius = Double

Radius of Orbit at the current Position

type Speed = Double

Speed is an alias for Double

var_G :: GravConst

var_G [Nm^2/kg^2] is the Gravitation constant in newton meter squared over kilo gramms squared

semiMajor :: Orbit Body -> Double

semiMajor calculates the semi major axis of an orbit.

v :: Orbit Body -> Radius -> Speed

v takes an orbit and a radius (from the center of the centerBody) and calculates the orbital speed at that position.

v_e :: Body -> Speed

v_e calculates the escape velocity of that body.

hohmann :: Orbit Body -> Orbit Body -> (Double, Double)

hohmann takes two orbits around the same centerBody. It calculates both (v1 and v2 ) delta V changes for a hohmann transfair.