System.KSP.Universe

Description

Documentation

kMoho :: Body

kMoho is the definition of Moho

kGilly :: Body

kGilly is the definition of Gilly

kEve :: Body

kEve is the definition of Eve

kMun :: Body

kMun is the definition of Mun

kMinmus :: Body

kMinmus is the definition of Minmus

kKerbin :: Body

kKerbin is the definition of Kerbin

kIke :: Body

kIke is the definition of Ike

kDres :: Body

kDres is the definition of Dres

kDuna :: Body

kDuna is the definition of Duna

kLaythe :: Body

kLaythe is the definition of Laythe

kVall :: Body

kVall is the definition of Vall

kTylo :: Body

kTylo is the definition of Tylo

kBop :: Body

kBop is the definition of Bop

kPol :: Body

kPol is the definition of Pol

kJool :: Body

kJool is the definition of Jool

kEeloo :: Body

kEeloo is the definition of Eeloo

kKerbol :: Body

kKerbol is the definition of Kerbol

kGillyOrbit :: Orbit Body

kGillyOrbit is the definition of the Orbit of Gilly

kMunOrbit :: Orbit Body

kMunOrbit is the definition of the Orbit of Mun

kMinmusOrbit :: Orbit Body

kMinmusOrbit is the definition of the Orbit of Minmus

kIkeOrbit :: Orbit Body

kIkeOrbit is the definition of the Orbit of Ike

kLaytheOrbit :: Orbit Body

kLaytheOrbit is the definition of the Orbit of Laythe

kVallOrbit :: Orbit Body

kVallOrbit is the definition of the Orbit of Vall

kTyloOrbit :: Orbit Body

kTyloOrbit is the definition of the Orbit of Tylo

kBopOrbit :: Orbit Body

kBopOrbit is the definition of the Orbit of Bop

kPolOrbit :: Orbit Body

kPolOrbit is the definition of the Orbit of Pol

kMohoOrbit :: Orbit Body

kMohoOrbit is the definition of the Orbit of Moho

kEveOrbit :: Orbit Body

kEveOrbit is the definition of the Orbit of Eve

kKerbinOrbit :: Orbit Body

kKerbinOrbit is the definition of the Orbit of Kerbin

kDunaOrbit :: Orbit Body

kDunaOrbit is the definition of the Orbit of Duna

kDresOrbit :: Orbit Body

kDresOrbit is the definition of the Orbit of Dres

kJoolOrbit :: Orbit Body

kJoolOrbit is the definition of the Orbit of Jool

kEelooOrbit :: Orbit Body

kEelooOrbit is the definition of the Orbit of Eeloo